CETERRA-18 implemented abstract GlobalDisplay class
