+++ Assets/Code/Scripts/Domains/GlobalDisplay/EconomyDisplay.cs
@@ -1,0 +1,5 @@
+public class EconomyDisplay : GlobalDisplay {
+    public EconomyDisplay(int currentValue) {
+        this.CurrentValue = currentValue;
+    }
+}
