+++ Assets/Code/Scripts/Domains/GlobalDisplay/EnvironmentDisplay.cs
@@ -1,0 +1,5 @@
+public class EnvironmentDisplay : GlobalDisplay {
+    public EnvironmentDisplay(int currentValue) {
+        this.CurrentValue = currentValue;
+    }
+}
