+++ Assets/Code/Scripts/Domains/GlobalDisplay/EnvironmentDisplay.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c1aee6123b856cc438628e88fff4dc56
