+++ Assets/Code/Scripts/Domains/GlobalDisplay/FondsDisplay.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 764b716340bbdef419b50f91f65e12bb
