+++ Assets/Code/Scripts/Domains/GlobalDisplay/ResourceDisplay.cs
@@ -1,0 +1,16 @@
+using System;
+
+public class ResourceDisplay : GlobalDisplay {
+    public ResourceDisplay(int currentValue) {
+        this.CurrentValue = currentValue;
+    }
+
+    public override void SubtractCurrentValue(int subtract) {
+        if (this.CurrentValue - subtract < this.MinValue) {
+            throw new Exception("Subtraction exceeds minimum value " + this.MinValue);
+        }
+        else {
+            this.CurrentValue -= subtract;
+        }
+    }
+}
