+++ Assets/Code/Scripts/Domains/GlobalDisplay/SocietyDisplay.cs
@@ -1,0 +1,5 @@
+public class SocietyDisplay : GlobalDisplay {
+    public SocietyDisplay(int currentValue) {
+        this.CurrentValue = currentValue;
+    }
+}
