+++ Assets/Code/Scripts/Domains/GlobalDisplay/SocietyDisplay.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: de6bb27078a426e43af354cb3d1469cd
