--- Assets/Code/Scripts/EconomyDisplay.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e8df3c1c956c1bd48acadb5a1381b101
