--- Assets/Code/Scripts/EnvironmentDisplay.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e491e360fc1ab4d48a3f0e0c62c7da64
