--- Assets/Code/Scripts/FondsDisplay.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3aa6c5b6627f44141876e567501ac78c
