--- Assets/Code/Scripts/GlobalDisplay.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9431802d62a43f04aa3e44dee4ebe9b1
