--- Assets/Code/Scripts/ResourcesDIsplay.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7485b8053d155cb4cae91f57308e86b1
