--- Assets/Code/Scripts/SocietyDisplay.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d8c1a6260f21315439845e47206a8090
